+++ src/components/menu/menu.styles.ts
@@ -11,4 +11,8 @@ export default css`
   .menu {
     padding: var(--sl-spacing-x-small) 0;
   }
+
+  ::slotted(sl-divider) {
+    --spacing: var(--sl-spacing-x-small);
+  }
 `;